requirements.txt
streamlit
google-search-results
langchain
openai
langchain-openai
python-dotenv
openpyxl
langchain-experimental
numexpr
beautifulsoup4